internal: Refactor common bits from envoy/listeners
Updates #1898
