+++ internal/envoy/v2/listener.go
@@ -15,7 +15,6 @@ package v2
 
 import (
 	"fmt"
-	"log"
 	"sort"
 	"strings"
 	"time"
@@ -37,65 +36,6 @@ import (
 	"github.com/projectcontour/contour/internal/sorter"
 	"github.com/projectcontour/contour/internal/timeout"
 )
-
-type HTTPVersionType = http.HttpConnectionManager_CodecType
-
-const (
-	HTTPVersionAuto HTTPVersionType = http.HttpConnectionManager_AUTO
-	HTTPVersion1    HTTPVersionType = http.HttpConnectionManager_HTTP1
-	HTTPVersion2    HTTPVersionType = http.HttpConnectionManager_HTTP2
-	HTTPVersion3    HTTPVersionType = http.HttpConnectionManager_HTTP3
-)
-
-// ProtoNamesForVersions returns the slice of ALPN protocol names for the give HTTP versions.
-func ProtoNamesForVersions(versions ...HTTPVersionType) []string {
-	protocols := map[HTTPVersionType]string{
-		HTTPVersion1: "http/1.1",
-		HTTPVersion2: "h2",
-		HTTPVersion3: "",
-	}
-	defaultVersions := []string{"h2", "http/1.1"}
-	wantedVersions := map[HTTPVersionType]struct{}{}
-
-	if versions == nil {
-		return defaultVersions
-	}
-
-	for _, v := range versions {
-		wantedVersions[v] = struct{}{}
-	}
-
-	var alpn []string
-
-	// Check for versions in preference order.
-	for _, v := range []HTTPVersionType{HTTPVersionAuto, HTTPVersion2, HTTPVersion1} {
-		if _, ok := wantedVersions[v]; ok {
-			if v == HTTPVersionAuto {
-				return defaultVersions
-			}
-
-			log.Printf("wanted %d -> %s", v, protocols[v])
-			alpn = append(alpn, protocols[v])
-		}
-	}
-
-	return alpn
-}
-
-// CodecForVersions determines a single Envoy HTTP codec constant
-// that support all the given HTTP protocol versions.
-func CodecForVersions(versions ...HTTPVersionType) HTTPVersionType {
-	switch len(versions) {
-	case 1:
-		return versions[0]
-	case 0:
-		// Default is to autodetect.
-		return HTTPVersionAuto
-	default:
-		// If more than one version is allowed, autodetect and let ALPN sort it out.
-		return HTTPVersionAuto
-	}
-}
 
 // TLSInspector returns a new TLS inspector listener filter.
 func TLSInspector() *envoy_api_v2_listener.ListenerFilter {
@@ -140,7 +80,7 @@ type httpConnectionManagerBuilder struct {
 	maxConnectionDuration         timeout.Setting
 	connectionShutdownGracePeriod timeout.Setting
 	filters                       []*http.HttpFilter
-	codec                         HTTPVersionType // Note the zero value is AUTO, which is the default we want.
+	codec                         envoy.HTTPVersionType // Note the zero value is AUTO, which is the default we want.
 }
 
 // RouteConfigName sets the name of the RDS element that contains
@@ -161,7 +101,7 @@ func (b *httpConnectionManagerBuilder) MetricsPrefix(prefix string) *httpConnect
 }
 
 // Codec sets the HTTP codec for the manager. The default is AUTO.
-func (b *httpConnectionManagerBuilder) Codec(codecType HTTPVersionType) *httpConnectionManagerBuilder {
+func (b *httpConnectionManagerBuilder) Codec(codecType envoy.HTTPVersionType) *httpConnectionManagerBuilder {
 	b.codec = codecType
 	return b
 }
